chore: rename fetch_transaction script
